Add option to create subfolder during export operations

+++ Source/AssetRipper.GUI.Web/GameFileLoader.cs
@@ -49,27 +49,39 @@ public static class GameFileLoader
 		GameData = ExportHandler.LoadAndProcess(paths);
 	}
 
-	public static void ExportUnityProject(string path)
+	public static void ExportUnityProject(string path, bool createSubfolder)
 	{
+		path = AdjustPathForSubfolder(path, createSubfolder);
+
 		if (IsLoaded && IsValidExportDirectory(path))
 		{
-			Directory.Delete(path, true);
+			if (!createSubfolder)
+			{
+				Directory.Delete(path, true);
+			}
+
 			Directory.CreateDirectory(path);
 			ExportHandler.Export(GameData, path);
 		}
 	}
 
-	public static void ExportPrimaryContent(string path)
+	public static void ExportPrimaryContent(string path, bool createSubfolder)
 	{
+		path = AdjustPathForSubfolder(path, createSubfolder);
+
 		if (IsLoaded && IsValidExportDirectory(path))
 		{
-			Directory.Delete(path, true);
+			if (!createSubfolder)
+			{
+				Directory.Delete(path, true);
+			}
+
 			Directory.CreateDirectory(path);
-			Logger.Info(LogCategory.Export, "Starting export");
+			Logger.Info(LogCategory.Export, "Starting primary content export");
 			Logger.Info(LogCategory.Export, $"Attempting to export assets to {path}...");
 			Settings.ExportRootPath = path;
 			PrimaryContentExporter.CreateDefault(GameData).Export(GameBundle, Settings, LocalFileSystem.Instance);
-			Logger.Info(LogCategory.Export, "Finished exporting assets");
+			Logger.Info(LogCategory.Export, "Finished exporting primary content.");
 		}
 	}
 
@@ -78,6 +90,17 @@ public static class GameFileLoader
 		LibraryConfiguration settings = new();
 		settings.LoadFromDefaultPath();
 		return settings;
+	}
+
+	private static string AdjustPathForSubfolder(string path, bool createSubfolder)
+	{
+		if (createSubfolder)
+		{
+			string timestamp = DateTime.UtcNow.ToString("yyyyMMdd_HHmmss");
+			string subfolder = $"AssetRipper_export_{timestamp}";
+			return Path.Combine(path, subfolder);
+		}
+		return path;
 	}
 
 	private static bool IsValidExportDirectory(string path)
